% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{boot_ICC}
\alias{boot_ICC}
\title{Generates a bootstrap sample and estimates the ICC}
\usage{
boot_ICC(X, nt, Bmat, indB)
}
\arguments{
\item{X}{Data frame with the pairwise distances}

\item{nt}{Data frame with the number of trips by subject}

\item{Bmat}{Matrix with subjects identifiers in the resamples}

\item{indB}{Column in Bmat that geenrates the resample}
}
\value{
Data frame with the estimates of the ICC (r), the subjects' mean sum-of-squares (MSA), the between-subjects variance (sb), the total variance (st), and the within-subjects variance (se).
}
\description{
Generates a bootstrap sample and estimates the ICC
}
\keyword{internal}
