% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dist_mat}
\alias{dist_mat}
\title{Creates a matrix with the pairwise distances}
\usage{
dist_mat(dataD, nt)
}
\arguments{
\item{dataD}{Data frame created with \code{"genHD"} function.}

\item{nt}{Data frame with the number of trips by subject}
}
\value{
A list with two components:
\itemize{
  \item *D*. A matrix with the pairwise distances
  \item *data*. A data frame with the pairwise distances
}
}
\description{
Creates a matrix with the pairwise distances
}
\keyword{internal}
