% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FD_data}
\alias{FD_data}
\title{Creates a data frame with the discrete Fréchet distance between two trajectories}
\usage{
FD_data(data, i, j)
}
\arguments{
\item{data}{A list created with \code{tr_gen} function}

\item{i}{Position in the data of the first trajectory}

\item{j}{Position in the data of the second trajectory}
}
\value{
A data frame with the subjects and trips identifiers and the discrete Fréchet distance
}
\description{
Creates a data frame with the discrete Fréchet distance between two trajectories
}
\keyword{internal}
