% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_data}
\alias{get_data}
\title{Preprocessing of data}
\usage{
get_data(data, ID, trip, LON, LAT, time)
}
\arguments{
\item{data}{A data frame with the steps of trajectories. It is assumed the time between steps is uniform. It must contain contain at least five columns: subject identifier, trip identifier, latitude, longitude, and time of the reading.}

\item{ID}{Character string indicating the name of the subjects column in the dataset.}

\item{trip}{Character string indicating the trip column in the dataset.}

\item{LON}{Numeric. Longitude readings.}

\item{LAT}{Numeric. Latitude readings.}

\item{time}{Numeric. Time of the readings.}
}
\value{
A data frame
}
\description{
Preprocessing of data
}
\keyword{internal}
