% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gull_data}
\alias{gull_data}
\title{Gull data}
\format{
A data frame containing  90 trajectories
\describe{
  \item{ID}{Subject identifier}
  \item{trip}{Trip identifier}
  \item{LONG}{Longitude}
  \item{LAT}{Latitude}
  \item{triptime}{Time in seconds when the locations were obtained}
}
}
\usage{
gull_data
}
\description{
A data frame with sample of 90 gull trajectories.
}
\keyword{datasets}
