% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{HD_data}
\alias{HD_data}
\title{Creates a data frame with the Hausdorff distance between two trajectories}
\usage{
HD_data(data, i, j, q)
}
\arguments{
\item{data}{A list created with \code{"tr_gen"} function}

\item{i}{Position in the data of the first trajectory}

\item{j}{Position in the data of the second trajectory}

\item{q}{Quantile for the extended Hausdorff distance. Default value q=1 uses the maximum that leads to classical Hausdorff distance.}
}
\value{
A data frame with the subjects and trips identifiers and the extended Hausdorff distance
}
\description{
Creates a data frame with the Hausdorff distance between two trajectories
}
\keyword{internal}
