% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ICC}
\alias{ICC}
\title{Computes the intraclass correlation coefficient (ICC) using a matrix of distances.}
\usage{
ICC(X, nt)
}
\arguments{
\item{X}{Matrix with the pairwise distances.}

\item{nt}{Data frame with the number of trips by subject}
}
\value{
Data frame with the estimates of the ICC (r), the subjects' mean sum-of-squares (MSA), the between-subjects variance (sb), the total variance (st), and the within-subjects variance (se).
}
\description{
Computes the intraclass correlation coefficient (ICC) using a matrix of distances.
}
\details{
The intraclass correlation coefficient is estimated using the distance matrix among trajectories.
}
