% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ntrips}
\alias{ntrips}
\title{Computes the number of trips by subject}
\usage{
ntrips(data)
}
\arguments{
\item{data}{A data frame with the steps of trajectories. It is assumed the time between steps is uniform. It must contain contain at least five columns: subject identifier, trip identifier, latitude, longitude, and time of the reading.}
}
\value{
A data frame with the number of trips by subject.
}
\description{
Computes the number of trips by subject
}
\keyword{internal}
