% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.iccTraj}
\alias{print.iccTraj}
\title{Prints the ICC}
\usage{
\method{print}{iccTraj}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"iccTraj"}}
}
\value{
The ICC estimate.
}
\description{
Prints the ICC
}
\keyword{internal}
